feat: added owned extensions and did some refactoring.
